+++ internal/mods/system/dal/entity/ent/permissionresource/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int64) predicate.PermissionResource {
+func ID(id int) predicate.PermissionResource {
 	return predicate.PermissionResource(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int64) predicate.PermissionResource {
+func IDEQ(id int) predicate.PermissionResource {
 	return predicate.PermissionResource(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int64) predicate.PermissionResource {
+func IDNEQ(id int) predicate.PermissionResource {
 	return predicate.PermissionResource(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int64) predicate.PermissionResource {
+func IDIn(ids ...int) predicate.PermissionResource {
 	return predicate.PermissionResource(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int64) predicate.PermissionResource {
+func IDNotIn(ids ...int) predicate.PermissionResource {
 	return predicate.PermissionResource(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int64) predicate.PermissionResource {
+func IDGT(id int) predicate.PermissionResource {
 	return predicate.PermissionResource(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int64) predicate.PermissionResource {
+func IDGTE(id int) predicate.PermissionResource {
 	return predicate.PermissionResource(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int64) predicate.PermissionResource {
+func IDLT(id int) predicate.PermissionResource {
 	return predicate.PermissionResource(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int64) predicate.PermissionResource {
+func IDLTE(id int) predicate.PermissionResource {
 	return predicate.PermissionResource(sql.FieldLTE(FieldID, id))
 }
