+++ internal/mods/system/dal/entity/ent/rolepermission/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int64) predicate.RolePermission {
+func ID(id int) predicate.RolePermission {
 	return predicate.RolePermission(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int64) predicate.RolePermission {
+func IDEQ(id int) predicate.RolePermission {
 	return predicate.RolePermission(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int64) predicate.RolePermission {
+func IDNEQ(id int) predicate.RolePermission {
 	return predicate.RolePermission(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int64) predicate.RolePermission {
+func IDIn(ids ...int) predicate.RolePermission {
 	return predicate.RolePermission(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int64) predicate.RolePermission {
+func IDNotIn(ids ...int) predicate.RolePermission {
 	return predicate.RolePermission(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int64) predicate.RolePermission {
+func IDGT(id int) predicate.RolePermission {
 	return predicate.RolePermission(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int64) predicate.RolePermission {
+func IDGTE(id int) predicate.RolePermission {
 	return predicate.RolePermission(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int64) predicate.RolePermission {
+func IDLT(id int) predicate.RolePermission {
 	return predicate.RolePermission(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int64) predicate.RolePermission {
+func IDLTE(id int) predicate.RolePermission {
 	return predicate.RolePermission(sql.FieldLTE(FieldID, id))
 }
