+++ internal/mods/system/dal/entity/ent/userposition/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int64) predicate.UserPosition {
+func ID(id int) predicate.UserPosition {
 	return predicate.UserPosition(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int64) predicate.UserPosition {
+func IDEQ(id int) predicate.UserPosition {
 	return predicate.UserPosition(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int64) predicate.UserPosition {
+func IDNEQ(id int) predicate.UserPosition {
 	return predicate.UserPosition(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int64) predicate.UserPosition {
+func IDIn(ids ...int) predicate.UserPosition {
 	return predicate.UserPosition(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int64) predicate.UserPosition {
+func IDNotIn(ids ...int) predicate.UserPosition {
 	return predicate.UserPosition(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int64) predicate.UserPosition {
+func IDGT(id int) predicate.UserPosition {
 	return predicate.UserPosition(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int64) predicate.UserPosition {
+func IDGTE(id int) predicate.UserPosition {
 	return predicate.UserPosition(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int64) predicate.UserPosition {
+func IDLT(id int) predicate.UserPosition {
 	return predicate.UserPosition(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int64) predicate.UserPosition {
+func IDLTE(id int) predicate.UserPosition {
 	return predicate.UserPosition(sql.FieldLTE(FieldID, id))
 }
