+++ internal/mods/system/dal/entity/ent/userrole.go
@@ -17,8 +17,7 @@ import (
 type UserRole struct {
 	config `json:"-"`
 	// ID of the ent.
-	// field.primary_key.comment
-	ID int64 `json:"id,omitempty"`
+	ID int `json:"id,omitempty"`
 	// field.foreign_key.comment
 	UserID int64 `json:"user_id,omitempty"`
 	// field.foreign_key.comment
@@ -89,7 +88,7 @@ func (ur *UserRole) assignValues(columns []string, values []any) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			ur.ID = int64(value.Int64)
+			ur.ID = int(value.Int64)
 		case userrole.FieldUserID:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field user_id", values[i])
